fix: move DSP volume control methods to the base class, add abstract methods for parameter get / set
